Add NYSE "holiday" on 2018-12-05
Both [NYSE] and [NASDAQ] were closed on December 5 for the National Day
of Mourning for President George H. W. Bush.

I'm not sure if the term "Market Holiday" is appropriate (NY Times and
Kiplinger did use that term, NYSE did not), but since fincal doesn't
have a notion of non-holiday market closures, this commit adds it to
`.holidays`.

[NASDAQ]: https://www.nasdaq.com/article/is-the-stock-market-closed-for-president-bushs-funeral-cm1064501
[NYSE]: https://www.nyse.com/trader-update/history#110000115869

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>