aider: Fixed the error by replacing undefined variables with values from the `data` dictionary in the `Venue` class constructor.
# Aider chat conversation:

USER: Error
Traceback (most recent call last):
  File "C:\Repos\fincal\tests\test_venue.py", line 7, in setUp
    self.venue = Venue('../exchanges/nyse.yml')
                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Repos\fincal\exchanges\venue.py", line 40, in __init__
    self.partial_trading_days = partial_trading_days
                                ^^^^^^^^^^^^^^^^^^^^
NameError: name 'partial_trading_days' is not defined

ASSISTANT: The error is due to the use of undefined variables `partial_trading_days` and `non_trading_days` in the `Venue` class constructor. These variables are not defined in the scope of the `__init__` method.

The `Venue` class constructor is trying to assign the values of these undefined variables to `self.partial_trading_days` and `self.irregular_non_trading_days` respectively.

To fix this, we need to replace these undefined variables with the corresponding values from the `data` dictionary that is loaded from the yaml file.

The files to modify are:

1. `exchanges\venue.py`

Here are the changes:

```python
exchanges\venue.py
<<<<<<< HEAD
        self.partial_trading_days = partial_trading_days
        self.irregular_non_trading_days = non_trading_days
=======
        self.partial_trading_days = data['partial_trading_days']
        self.irregular_non_trading_days = data['non_trading_days']
>>>>>>> updated
```

This change will replace the undefined variables with the corresponding values from the `data` dictionary.