aider: Fixed the error by replacing the undefined variables with the corresponding values from the `data` dictionary in the `Venue` class constructor.
# Aider chat conversation:

USER: Error
Traceback (most recent call last):
  File "C:\Repos\fincal\tests\test_venue.py", line 7, in setUp
    self.venue = Venue('../exchanges/nyse.yml')
                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Repos\fincal\exchanges\venue.py", line 42, in __init__
    self.data_provided_from_date = data_provided_from_date
                                   ^^^^^^^^^^^^^^^^^^^^^^^
NameError: name 'data_provided_from_date' is not defined

ASSISTANT: The error is due to the use of the undefined variable `data_provided_from_date` in the `Venue` class constructor. This variable is not defined in the scope of the `__init__` method.

The `Venue` class constructor is trying to assign the value of this undefined variable to `self.data_provided_from_date`.

To fix this, we need to replace this undefined variable with the corresponding value from the `data` dictionary that is loaded from the yaml file.

The files to modify are:

1. `exchanges\venue.py`

Here are the changes:

```python
exchanges\venue.py
<<<<<<< HEAD
        self.data_provided_from_date = data_provided_from_date
        self.data_provided_through_date = data_provided_through_date
=======
        self.data_provided_from_date = data['data_provided_from_date']
        self.data_provided_through_date = data['data_provided_through_date']
>>>>>>> updated
```

This change will replace the undefined variables with the corresponding values from the `data` dictionary.